+++ src/components/CheckboxGroup/CheckboxGroup.tsx
@@ -165,7 +165,7 @@ export const CheckboxGroup = chakra(
         <HelperErrorText
           id={`${id}-helperErrorText`}
           isInvalid={isInvalid}
-          renderText={showHelperInvalidText}
+          isRenderedText={showHelperInvalidText}
           text={footnote}
           __css={styles.helperErrorText}
         />